Return file read pointer to start when getting frame at time
The protobuf reader seeks through the file for the message at a time. The
file read pointer remains pointing to that location in the file. It's more intuitive
that if a user call get_simulation_frame_at_time, they can then call it again
to get any frame at a time after and before the previous time.